+++ src/main/kotlin/com/jordyma/blink/folder/repository/CustomFolderRepository.kt
@@ -3,7 +3,9 @@ package com.jordyma.blink.folder.repository
 import com.jordyma.blink.folder.entity.Folder
 import com.jordyma.blink.user.entity.User
 
-interface CustomFolderRepository: FolderRepository {
+interface CustomFolderRepository {
 
-    override fun findAllByUser(user: User): List<Folder>
+    fun findAllByUser(user: User): List<Folder>
+
+    fun delete(folder: Folder): Unit
 }
