+++ src/main/kotlin/com/jordyma/blink/folder/repository/FolderRepository.kt
@@ -4,11 +4,13 @@ import com.jordyma.blink.folder.entity.Folder
 import com.jordyma.blink.user.entity.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
+import org.springframework.stereotype.Repository
 
-interface FolderRepository : JpaRepository<Folder, Long> {
-    fun findByUser(user: User): List<Folder>
+@Repository
+interface FolderRepository : JpaRepository<Folder, Long>, CustomFolderRepository {
+//    fun findByUser(user: User): List<Folder>
 
-    fun findAllByUser(user: User): List<Folder>
+    override fun findAllByUser(user: User): List<Folder>
 
     @Query("select f from Folder f where f.user =:user and f.isUnclassified =true")
     fun findUnclassified(user: User): Folder?
@@ -18,4 +20,6 @@ interface FolderRepository : JpaRepository<Folder, Long> {
 
     @Query("select f from Folder f where f.user =:user and f.name =:name")
     fun findByName(name: String, user: User): Folder?
+
+    override fun delete(folder: Folder): Unit
 }
