+++ src/main/kotlin/com/jordyma/blink/keyword/service/KeywordService.kt
@@ -1,12 +1,9 @@
 package com.jordyma.blink.keyword.service
 
-import com.jordyma.blink.feed.dto.request.FeedUpdateReqDto
 import com.jordyma.blink.feed.entity.Feed
-import com.jordyma.blink.feed.repository.CustomFeedRepository
 import com.jordyma.blink.feed.repository.FeedRepository
 import com.jordyma.blink.keyword.entity.Keyword
 import com.jordyma.blink.keyword.repository.KeywordRepository
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
